model_worker/requirements.txt
# Message Queue
pika>=1.3.0

# Storage
minio>=7.2.0

# Database
psycopg2-binary>=2.9.10

# AI/ML - YOLO
ultralytics>=8.3.0

# Utilities
python-dotenv>=1.0.0
pillow>=11.0.0
numpy>=1.23.0,<2.0.0
